[INTERNAL] split transform-apijson-for-sdk script
To separate execution logic from standalone module it is split into 2
separate files.
